Refactor code to support a different configuration
The new configuration style is based on the idea of blocks and flows. On
one hand the user defines a list of blocks that are going to be used
later to define flows. A flow is just the sequence of blocks that the
messages traverse from the input flot to the output flow.